refactor: add condition and throw error when the player not found

+++ src/main/java/enzosdev/elifoot/service/PlayerService.java
@@ -62,6 +62,9 @@ public class PlayerService {
     }
 
     public void deletePlayerById(Long id){
+        if(!playerRepository.existsById(id)){
+            throw new PlayerNotFoundException("Player Not Found");
+        }
         playerRepository.deleteById(id);
     }
